Modify README, add 2 new recipe pages
Filled in the README with a brief project description. Added 2 new recipe pages,
their links in index.html, and filled in body of each page to match the format of the
first recipe page. Also altered the links on index.html to open in blank tab and added
noopener and noreferral. Added missing titles to each recipe page.